refactor: Path를 상위 패키지 move로 이동
