+++ src/main/java/domain/move/movement/PieceMovement.java
@@ -2,6 +2,7 @@ package domain.move.movement;
 
 import domain.board.Intersection;
 import domain.game.Side;
+import domain.move.Path;
 import java.util.List;
 
 public abstract class PieceMovement {
