+++ src/main/java/domain/move/rule/MoveRule.java
@@ -2,7 +2,7 @@ package domain.move.rule;
 
 import domain.board.Intersection;
 import domain.game.Side;
-import domain.move.movement.Path;
+import domain.move.Path;
 import domain.piece.AlivePieces;
 import java.util.List;
