+++ src/test/java/domain/piece/moverule/BasicRuleTest.java
@@ -8,7 +8,7 @@ import domain.move.rule.BasicRule;
 import domain.piece.AlivePieces;
 import domain.piece.Piece;
 import domain.piece.Soldier;
-import domain.move.movement.Path;
+import domain.move.Path;
 import java.util.List;
 import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
